authstack/pyproject.toml
[project]
name = "authstack"
version = "0.1.0"
description = "Django + DRF + SimpleJWT + Roles/Perms starter"
requires-python = ">=3.10"
dependencies = [
  "django==5.2.5",
  "djangorestframework==3.16.1",
  "djangorestframework-simplejwt==5.5.1",
  "django-cors-headers==4.7.0",
  "PyJWT==2.10.1",
]
